Initial commit
This commit adds the basic framework for interacting with mochad via
python. Right now it's basic and just contains functions for sending
a command and receiving the response. Eventually this will grow and
add support for basic protocol commands to abstract common operations
and not rely on a user knowing the X10 protocol.