Deduplicate the connection logic
With the introduction of the reconnect method on the controller class
the loop used to open a socket connection to the mochad daemon was
duplicated between there and the __init__() function. We only need to
maintain a single copy of this though. This commit addresses this by
creating a new private function, _connect(), which will establish open a
socket connection to the daemon and uses it in both __init__() and
reconnect().