Add travis configuration to repo
This commit adds configuration to the repo for running travis ci jobs.
Right now there is no ci on the repo which means everything is left up
to review. While there aren't any unit tests on the project yet we can
at least check style rules and that the docs build on proposed commits.
Unit tests will come in the future.